slack-bot/pyproject.toml
[tool.poetry]
name = "slack-bot"
version = "0.1.0"
description = ""
authors = [" "]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
slack-sdk = "^3.27.1"
slack-bolt = "^1.18.1"
httpx = "^0.27.0"
fastapi = "^0.104.1"
uvicorn = "^0.24.0.post1"
emoji = "^2.9.0"

pydantic = "^1.10.4"

google-cloud-secret-manager = "^2.16.3"
google-cloud-logging = "^3.8.0"

requests = "^2.31.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
